Rename: Auth-Client 위치 Shared 레이어로 변경
