+++ src/features/auth/login/index.ts
@@ -1,7 +1,2 @@
-// model
-export * from './model/auth-client'
-export * from './model/auth-interceptor'
-export * from './model/refresh-access-token'
-
 //ui
-export * from './ui/login-button'
+export * from './ui/login-button';
