+++ src/features/join-group/api/use-join-to-group.ts
@@ -1,6 +1,5 @@
 import { groupQuery } from "@/entities/group/api/group.query";
-import { post } from "@/features/auth/login";
-import { errorToast, successToast } from "@/shared/lib";
+import { errorToast, post, successToast } from "@/shared/lib";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { JOIN_GROUP_MESSAGES } from "../model/constants";
