+++ src/shared/lib/auth-client.ts
