+++ src/shared/lib/index.ts
@@ -1,4 +1,5 @@
+export * from './auth-client'
 export * from './create-image-handler'
 export * from './toast'
-export * from './utils'
 export * from './use-confetti'
+export * from './utils'
